+++ packages/ng/src/app/select/index.ts
@@ -1,0 +1,5 @@
+export { LuSelectModule } from './select.module';
+export { LuSelectPopover } from './select.popover.component';
+export { LuSelectOption } from './select.option.component';
+export { LuSelectDirective } from './select.directive';
+export { LuSelect } from './select.component';